backend/auth-service/requirements.txt
fastapi==0.110.0
uvicorn==0.27.0

# Security / JWT
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==3.2.2
python-multipart==0.0.6
pydantic[email]==2.6.0
pydantic-settings==2.2.1
python-dotenv==1.0.1

# OAuth2 / OIDC + Social login
authlib==1.3.0
itsdangerous==2.2.0

# Database (PostgreSQL) + migrations
SQLAlchemy==2.0.27
# Note: asyncpg wheels may not exist for very new Python versions on Windows.
# Our runtime is Docker (python:3.11-slim), so this is fine there.
asyncpg==0.29.0; python_version < "3.13"
alembic==1.13.1
psycopg[binary]==3.1.18

# Redis (sessions / rate-limit / blacklist)
redis==5.0.1

# MFA
pyotp==2.9.0
qrcode==7.4.2

# Metrics
prometheus-client==0.20.0

# Tests (dev)
pytest==8.0.2
pytest-asyncio==0.23.5
httpx==0.26.0
